# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CAB Booking System API",
    "version": "1.0.0",
    "description": "Aggregated OpenAPI snapshot for the current CAB Booking System repo state."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/pricing/estimate": {
      "post": { "tags": ["pricing"], "summary": "Estimate fare", "requestBody": { "$ref": "#/components/requestBodies/PricingEstimateRequest" }, "responses": { "200": { "description": "Estimate" } } }
    },
...
    "/api/eta/estimate": {
      "post": { "tags": ["eta"], "summary": "Estimate ETA", "requestBody": { "$ref": "#/components/requestBodies/EtaEstimateRequest" }, "responses": { "200": { "description": "ETA" } } }
    },
...
    "/api/realtime/drivers/{driverId}/location": {
      "get": {
        "tags": ["realtime"],
        "summary": "Get driver realtime location",
        "parameters": [{ "$ref": "#/components/parameters/DriverIdParam" }],
        "responses": { "200": { "description": "Driver location" } }
      }
    },
...
    "/api/realtime/rides/{rideId}/route": {
      "get": {
        "tags": ["realtime"],
        "summary": "Get ride route",
        "parameters": [{ "$ref": "#/components/parameters/RideIdParam" }],
        "responses": { "200": { "description": "Route" } }
      }
    }
  },
  "components": {
    "parameters": {
...
      "DriverIdParam": {
        "name": "driverId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "RideIdParam": {
        "name": "rideId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "requestBodies": {
...
      "PricingEstimateRequest": {
        "required": true,
        "content": { "application/json": { "schema": { "type": "object", "properties": { "distance_km": { "type": "number" }, "demand_index": { "type": "number" }, "supply_index": { "type": "number" }, "vehicleType": { "type": "string" } } } } }
      },
      "EtaEstimateRequest": {
        "required": true,
        "content": { "application/json": { "schema": { "type": "object", "properties": { "distance_km": { "type": "number" }, "traffic_level": { "type": "number" }, "avg_speed_kmh": { "type": "number" }, "weather_factor": { "type": "number" } } } } }
      }
    },
...
  }
}